% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_split.R
\name{dep_split}
\alias{dep_split}
\title{Test/train split}
\usage{
dep_split(data, p, seed = NULL)
}
\arguments{
\item{data}{Name and path of input file}

\item{p}{Training/Testing split percentage. Should be between 0 and 1.}

\item{seed}{Set the seed of R's random number generator}
}
\value{
The original dataframe with an additional \code{train} column. p*100\% of the KNM pairs of dependencies are assigned to the training set.
}
\description{
Test/train split
}
