% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignore_dep.R
\name{ignore_dep}
\alias{ignore_dep}
\title{ignore dependence}
\usage{
ignore_dep(KM_train, KM_test, KNM_train, KNM_test, unknown = NULL)
}
\arguments{
\item{KM_train}{Known matches from training dataset}

\item{KM_test}{Known matches from testing dataset}

\item{KNM_train}{Known non matches from training dataset}

\item{KNM_test}{Known non matches from testing dataset}

\item{unknown}{a data frame of scores from an unknown case. Column names should match the variable names set for scores.}
}
\value{
slr evaluation for ignore dependence method
}
\description{
ignore dependence
}
