% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indep_km_set.R
\name{indep_km_set}
\alias{indep_km_set}
\title{independent known match set}
\usage{
indep_km_set(KM_train, seed = NULL)
}
\arguments{
\item{KM_train}{Known matches from training dataset}

\item{seed}{Random seed}
}
\value{
independent KM set of the training data using \code{seed}
Returns an independent KM set of the training data using \code{seed}
}
\description{
independent known match set
}
