% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indep_knm_set.R
\name{indep_knm_set}
\alias{indep_knm_set}
\title{Independent known non match set}
\usage{
indep_knm_set(KNM_train, seed = NULL)
}
\arguments{
\item{KNM_train}{known non matches (KNM) from training dataset.}

\item{seed}{seed for R's random number generator.}
}
\value{
dataframe with same columns as \code{KNM_train} only containing an independent KM set of the training data.
}
\description{
This function returns an independent set of known non matches (KNM) using \code{seed}.
}
