% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_kde.R
\name{multiple_kde}
\alias{multiple_kde}
\title{Calculate the SLR by computing KDE of numerator and denominator on multiple independent sets and averaging the estimated densities.}
\usage{
multiple_kde(
  KM_train,
  KM_test,
  KNM_train,
  KNM_test,
  NUM_SETS = 10,
  unknown = NULL
)
}
\arguments{
\item{KM_train}{Known matches from training dataset}

\item{KM_test}{Known matches from testing dataset}

\item{KNM_train}{Known non matches from training dataset}

\item{KNM_test}{Known non matches from testing dataset}

\item{NUM_SETS}{number of sets to use}

\item{unknown}{a data frame of scores from an unknown case. Column names should match the variable names set for scores.}
}
\value{
slr evaluation for multiple kde method
}
\description{
Calculate the SLR by computing KDE of numerator and denominator on multiple independent sets and averaging the estimated densities.
}
