% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_thresh.R
\name{opt_thresh}
\alias{opt_thresh}
\title{optimal threshold}
\usage{
opt_thresh(KM_SLR_train, KNM_SLR_train)
}
\arguments{
\item{KM_SLR_train}{Known matches from training dataset}

\item{KNM_SLR_train}{Known non matches from training dataset}
}
\value{
threshold for SLR
}
\description{
optimal threshold
}
