% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_slr_roc.R
\name{plot_slr_roc}
\alias{plot_slr_roc}
\title{Plot ROC curve for selected method}
\usage{
plot_slr_roc(
  data,
  method = "AverageFeatures",
  num_runs = 200,
  alpha = 0.1,
  NUM_SETS = 10,
  seed = NULL
)
}
\arguments{
\item{data}{Dataset}

\item{method}{Method to calculate SLR options include c("IgnoreDependence", "StrictIndependentSet", "AverageFeatures", "MultipleKDE")}

\item{num_runs}{number of runs to run slr results}

\item{alpha}{transparency parameter from ggplot2}

\item{NUM_SETS}{number of sets to use for multiple kde method}

\item{seed}{Set the seed of R‘s random number generator}
}
\value{
ROC plot of selected method
}
\description{
Plot ROC curve for selected method
}
