% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoedata.R
\docType{data}
\name{shoedata}
\alias{shoedata}
\title{An example data set}
\format{
A data frame with 926 rows and 7 variables:
\describe{
\item{source1}{shoe identifier for the first shoe}
\item{dep1}{pair identifier for the first shoe}
\item{source2}{shoe identifier for the second shoe}
\item{dep2}{pair identifier for the second shoe}
\item{clique_size}{score; size of the maximal clique}
\item{med_dist_euc}{score; median Euclidean distance between overlapping pixels}
\item{input_overlap}{score; proportion of overlapping pixels in the aligned images}
}
}
\source{
Han (2022)
}
\usage{
shoedata
}
\description{
An example data set
}
\keyword{datasets}
