% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strict_indep_set.R
\name{strict_indep_set}
\alias{strict_indep_set}
\title{strict independent set}
\usage{
strict_indep_set(
  KM_train,
  KM_test,
  KNM_train,
  KNM_test,
  unknown = NULL,
  seed = NULL
)
}
\arguments{
\item{KM_train}{Known matches from training dataset}

\item{KM_test}{Known matches from testing dataset}

\item{KNM_train}{Known non matches from training dataset}

\item{KNM_test}{Known non matches from testing dataset}

\item{unknown}{Data frame of scores from an unknown case. Column names should match the variable names set for scores.}

\item{seed}{Random seed}
}
\value{
slr evaluation for independent set method
}
\description{
strict independent set
}
