% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_split.R
\name{dep_split}
\alias{dep_split}
\title{Test/train split}
\usage{
dep_split(data, p, seed = NULL)
}
\arguments{
\item{data}{data frame containing columns named \code{source1}, \code{source2}, \code{dep1}, and \code{dep2}. If \code{source1==source2}, this should indicate a known match (KM) pair, and if \code{source1!=source2}, this should indicate a known non-match (KNM) pair. The columns \code{dep1} and \code{dep2} should indicate dependence that you'd like the training/testing split to respect.}

\item{p}{training/testing split percentage. Should be between 0 and 1.}

\item{seed}{seed for R's random number generator.}
}
\value{
a data frame with the same columns as the input \code{data} data frame with an additional \code{train} column.
}
\description{
This function creates a column assigning rows to either the training or testing set. p*100\% of the KNM pairs of dependencies are assigned to the training set.
}
\examples{
head(dep_split(shoedata, .75))
}
