% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indep_km_set.R
\name{indep_km_set}
\alias{indep_km_set}
\title{Independent known match set}
\usage{
indep_km_set(KM_train, seed = NULL)
}
\arguments{
\item{KM_train}{known matches (KM) from training dataset.}

\item{seed}{seed for R's random number generator.}
}
\value{
dataframe with same columns as \code{KM_train} only containing an independent KM set of the training data.
}
\description{
This function returns an independent set of known matches (KM) using \code{seed}.
}
