% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indep_knm_set.R
\name{indep_knm_set}
\alias{indep_knm_set}
\title{independent known non match set}
\usage{
indep_knm_set(KNM_train, seed = NULL)
}
\arguments{
\item{KNM_train}{Known non matches from training dataset}

\item{seed}{Random seed}
}
\value{
independent KNM set of the data using \code{seed}
Given KNM_train, returns an independent KNM set of the data using \code{seed}
}
\description{
independent known non match set
}
