% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_slr_roc.R
\name{plot_slr_roc}
\alias{plot_slr_roc}
\title{Plot ROC curve for selected method}
\usage{
plot_slr_roc(
  data,
  method = "AverageFeatures",
  num_runs = 200,
  alpha = 0.1,
  NUM_SETS = 10,
  seed = NULL
)
}
\arguments{
\item{data}{data frame containing columns named \code{source1}, \code{source2}, \code{dep1},  \code{dep2}, and \code{train}. The \code{train} column should be a logical vector with \code{TRUE} denoting the training set and \code{FALSE} the test set. Any remaining columns must be scores. There may be up to five scores.}

\item{method}{string. Options are "IgnoreDependence", "StrictIndependentSet", "AverageFeatures", and "MultipleKDE".}

\item{num_runs}{number of times to run \code{slr_results}.}

\item{alpha}{transparency parameter from \code{ggplot2} package.}

\item{NUM_SETS}{number of sets to use for multiple KDE method.}

\item{seed}{seed of R‘s random number generator.}
}
\value{
plot of selected method.
}
\description{
This function returns a plot of the ROC curves produced by \code{num_runs} times of running \code{\link[=slr_results]{slr_results()}}. Note that this function only supports plotting one method at a time.
}
\examples{
# Set up
shoedata_split <- dep_split(shoedata, 0.75)

plot_slr_roc(shoedata_split, method = "StrictIndependentSet", num_runs = 3)
}
\seealso{
\code{\link[=ignore_dep]{ignore_dep()}} for "IgnoreDependence" method, \code{\link[=strict_indep_set]{strict_indep_set()}} for "StrictIndependentSet" method, \code{\link[=avg_features]{avg_features()}} for "AverageFeatures" method, and \code{\link[=multiple_kde]{multiple_kde()}} for "MultipleKDE" method.
}
