% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr_results.R
\name{slr_results}
\alias{slr_results}
\title{Results of one or more method of computing the SLR}
\usage{
slr_results(
  data,
  unknown = NULL,
  method = c("IgnoreDependence", "StrictIndependentSet", "AverageFeatures",
    "MultipleKDE"),
  NUM_SETS = 10,
  seed = NULL
)
}
\arguments{
\item{data}{Dataset}

\item{unknown}{a data frame of scores from an unknown case. Column names should match the variable names set for scores.}

\item{method}{Method to calculate SLR Method to calculate SLR options include c("IgnoreDependence", "StrictIndependentSet", "AverageFeatures", "MultipleKDE")}

\item{NUM_SETS}{number of sets to use for multiple kde method}

\item{seed}{Set the seed of R‘s random number generator}
}
\value{
slr results
}
\description{
Results of one or more method of computing the SLR
}
